issues-5: rename the artefactId, groupId and package name to the common used scheme of the new MagKit modules
