+++ .gitignore
@@ -3,6 +3,6 @@
 /.project
 /.classpath
 /.README.md.html
-/magnolia-vanity-url.iml
+/magkit-vanity-url.iml
 .checkstyle
