+++ src/main/resources/magkit-vanity-url/contentTypes/vanityUrl.yaml
