+++ src/main/resources/magkit-vanity-url/virtualUriMappings/vanityUrls.yaml
@@ -1,0 +1,1 @@
+class: de.ibmix.magkit.vanity.url.VirtualVanityUriMapping
