--- src/main/resources/magnolia-vanity-url/i18n/messages_de.properties
@@ -1,50 +1,0 @@
-###
-# #%L
-# magnolia-vanity-url Magnolia Module
-# %%
-# Copyright (C) 2013 - 2014 Aperto AG
-# %%
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# 
-#      http://www.apache.org/licenses/LICENSE-2.0
-# 
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# #L%
-###
-vanityUrls.app.label=Vanity urls
-vanityUrls.app.preview.label=Vorschau
-vanityUrls.detail.label=Vanity url
-
-vanityUrls.browser.actionbar.sections.item.label=Vanity url
-vanityUrls.browser.actions.addItem.label=Url hinzufügen
-vanityUrls.browser.actions.editItem.label=Url bearbeiten
-
-vanityUrls.browser.actions.preview.label=Vorschau
-vanityUrls.browser.actions.viewCode.label=Anzeige QR code
-vanityUrls.browser.views.vanityUrl.label=Vanity url
-vanityUrls.browser.views.site.label=Site
-vanityUrls.browser.views.link.label=Ziellink
-vanityUrls.browser.views.activationStatus.label=Status
-
-vanityUrls.vanityUrl.label=Vanity url
-vanityUrls.vanityUrl.description=Vanity url startet mit einem Slash, z.B. /xmas
-vanityUrls.site.label=Site
-vanityUrls.site.description=Ohne Multisite-Support von Magnolia 'fallback' ist der Default.
-vanityUrls.link.label=Ziel
-vanityUrls.link.description=Wähle eine Seite oder füge eine externe Url ein, welche mit 'http://' or 'https://' beginnen sollte.
-vanityUrls.linkSuffix.label=Link-Suffix
-vanityUrls.linkSuffix.description=Optionales Suffix für die Linkerzeugung. Anker und Parameter wie '?query=app' or '#chapter1' werden unterstützt.
-vanityUrls.type.label=Weiterleitung
-vanityUrls.type.description=Typ der Weiterleitung nach Aufruf der Vanity url.
-vanityUrls.type.options.301=301 Permanent
-vanityUrls.type.options.302=302 Temporary
-vanityUrls.type.options.forward=Forward
-
-vanityUrl.errorMessage.notUnique=Vanity url ist nicht eindeutig. Bitte ändere die Vanity url oder die Site.
-vanityUrl.errorMessage.invalid=Vanity url muss mit einem Slash beginnen.
