--- src/main/resources/magnolia-vanity-url/i18n/messages_en.properties
@@ -1,50 +1,0 @@
-###
-# #%L
-# magnolia-vanity-url Magnolia Module
-# %%
-# Copyright (C) 2013 - 2014 Aperto AG
-# %%
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# 
-#      http://www.apache.org/licenses/LICENSE-2.0
-# 
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# #L%
-###
-vanityUrls.app.label=Vanity urls
-vanityUrls.app.preview.label=Preview
-vanityUrls.detail.label=Vanity url
-
-vanityUrls.browser.actionbar.sections.item.label=Vanity url
-vanityUrls.browser.actions.addItem.label=Add vanity url
-vanityUrla.browser.actions.editItem.label=Edit vanity url
-
-vanityUrls.browser.actions.preview.label=Preview target
-vanityUrls.browser.actions.viewCode.label=View QR code
-vanityUrls.browser.views.vanityUrl.label=Vanity url
-vanityUrls.browser.views.site.label=Site
-vanityUrls.browser.views.link.label=Target link
-vanityUrls.browser.views.activationStatus.label=Status
-
-vanityUrls.vanityUrl.label=Vanity url
-vanityUrls.vanityUrl.description=Vanity url starts with slash, e.g. /xmas
-vanityUrls.site.label=Site
-vanityUrls.site.description=Without multi site support of Magnolia 'fallback' is the default.
-vanityUrls.link.label=Destination
-vanityUrls.link.description=Browse in your pages or add an external link, which should start with 'http://' or 'https://'.
-vanityUrls.linkSuffix.label=Link suffix
-vanityUrls.linkSuffix.description=Optional suffix for the link creation. You can provide anchor or parameters like '?query=app' or '#chapter1'.
-vanityUrls.type.label=Redirect
-vanityUrls.type.description=Type of the redirect after request the vanity url.
-vanityUrls.type.options.301=301 Permanent
-vanityUrls.type.options.302=302 Temporary
-vanityUrls.type.options.forward=Forward
-
-vanityUrl.errorMessage.notUnique=Vanity url is not unique. Please change vanity url or site configuration.
-vanityUrl.errorMessage.invalid=Vanity url must start with a slash.
