--- src/main/resources/magnolia-vanity-url/virtualUriMappings/vanityUrls.yaml
@@ -1,1 +1,0 @@
-class: com.aperto.magnolia.vanity.VirtualVanityUriMapping
