+++ src/main/resources/mgnl-bootstrap/magkit-vanity-url/config.modules.magnolia-vanity-url.config.excludes.xml
