--- src/main/resources/mgnl-nodetypes/magnolia-vanity-url-nodetypes.xml
@@ -1,26 +1,0 @@
-<!--
-  #%L
-  magnolia-vanity-url Magnolia Module
-  %%
-  Copyright (C) 2013 - 2014 Aperto AG
-  %%
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  
-       http://www.apache.org/licenses/LICENSE-2.0
-  
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  #L%
-  -->
-<nodeTypes xmlns:mgnl="http://www.magnolia.info/jcr/mgnl">
-    <nodeType name="mgnl:vanityUrl" isMixin="false" hasOrderableChildNodes="true" primaryItemName="">
-        <supertypes>
-            <supertype>mgnl:content</supertype>
-        </supertypes>
-    </nodeType>
-</nodeTypes>
